refactor: rename rate limiter
