+++ src/measurement/runner.ts
@@ -9,7 +9,7 @@ import { getMetricsAgent, type MetricsAgent } from '../lib/metrics.js';
 import type { MeasurementStore } from './store.js';
 import { getMeasurementStore } from './store.js';
 import type { MeasurementRequest, MeasurementResultMessage, MeasurementProgressMessage, UserRequest } from './types.js';
-import { rateLimit } from '../lib/ratelimiter.js';
+import { rateLimit } from '../lib/rate-limiter.js';
 
 export class MeasurementRunner {
 	constructor (
